Changed int to float for non-integer POL top up

+++ tests.py
@@ -5,7 +5,7 @@ from dataclass import Rawdata
 def bosDone_test(event: Rawdata):
     pass
 def pol_test(event: Rawdata):
-    if event.purpose == 'POL' and int(event.fuelAmount) == 0:
+    if event.purpose == 'POL' and float(event.fuelAmount) == 0:
         print('Error in {}, no fuel topped up at POL'.format(event.id))
 def chronoCheckVehicle_test(event: Rawdata, prev):
     if prev is not None and event.timeStart < prev.timeEnd:
